Collaborative-music-playing-system-main/backend/requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyJWT==2.8.0
requests==2.31.0
cryptography==42.0.5
python-jose[cryptography]==3.3.0
gunicorn==21.2.0
eventlet==0.35.2
